refactor: merge nats-interceptor into common
