+++ libs/common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './common.module';
 export * from './common.service';
+export * from './nats.interceptor';
